pollyvent/yvent/requirements.txt
Pillow>=9.0
qrcode>=7.3
